CMake: Handle dependencies through an interface.
This moves all the dependency finding, linking, (and workarounds), to
one place.

In addition:
- SDL2's minimum version was bumped to 2.0.12, this is the first version
where we are guaranteed that targets are defined. Moreover, earlier
versions had issues with the way include paths were set up on targets.
This allows us to drop the Find module.
- Avoid creating target aliases with upstream names, generator
expressions or if/else are prefered.
- Prefer linking to GL and GLU targets instead of variables.
- Ensure X11 GL is not picked over the system's one.